Add upgrade controls to reaplce artwork
