+++ apps/web/src/modules/create-proposal/components/TransactionForm/Upgrade/index.ts
@@ -1,0 +1,2 @@
+export * from './UpgradeInProgress'
+export * from './UpgradeRequired'
